refactor: WithAssociatedFile now creates file only if it doesn't exist

+++ bot/WithAssociatedFile.py
@@ -9,12 +9,13 @@ def WithAssociatedFile(cls):
     class Wrapper(cls):
         def __init__(self, *args):
             self.__associated_file_path = None
-            self.create_associated_file()
+            self.create_associated_file_if_not_exists()
 
-        def create_associated_file(self):
+        def create_associated_file_if_not_exists(self):
             self.__associated_file_path = os.path.join(Paths.BotGeneratedFolder, self.get_id() + '.py')
-            with open(self.__associated_file_path, 'w') as file:
-                pass
+            if not os.path.exists(self.__associated_file_path):
+                with open(self.__associated_file_path, 'w') as file:
+                    pass
 
         def get_associated_file(self):
             return self.__associated_file_path
